+++ SEMI_MZ/src/mz/report/model/dao/ReportDao.java
@@ -20,10 +20,10 @@ import mz.report.model.vo.Report;
 
 public class ReportDao {
 	private Properties prop = new Properties();
-	
+
 	public ReportDao() {
 		String fileName = MemberDao.class.getResource("/sql/report/report-mapper.xml").getPath();
-		
+
 		try {
 			prop.loadFromXML(new FileInputStream(fileName));
 		} catch (InvalidPropertiesFormatException e) {
@@ -34,13 +34,12 @@ public class ReportDao {
 			e.printStackTrace();
 		}
 	}
-	
+
 //------------------------------ select 구간 -------------------------------
-	
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public int reportCount(Connection conn, String search, String keyword) {
-		
+
 		int result = 0;
 
 		ResultSet rset = null;
@@ -50,13 +49,13 @@ public class ReportDao {
 
 		if (search.equals("userId")) {
 			sql = sql.replaceAll("##", "USER_ID");
-		} else if(search.equals("receveId")) {
+		} else if (search.equals("receveId")) {
 			sql = sql.replaceAll("##", "RECEIVE_ID");
 		} else {
 			sql = sql.replaceAll("##", "REPORT_TITLE");
 		}
-		
-		// like문에 들어갈 문자열 혹시 모를 _ 와 % escape 처리 
+
+		// like문에 들어갈 문자열 혹시 모를 _ 와 % escape 처리
 		String like = keyword.replaceAll("%", "\\%");
 		like = like.replaceAll("_", "\\_");
 		like = "%" + like + "%";
@@ -78,14 +77,14 @@ public class ReportDao {
 			close(rset);
 			close(pstmt);
 		}
-		
+
 		return result;
 	}
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public ArrayList<Report> selectReportPart(Connection conn, int page, String search, String keyword) {
-		
-		ArrayList<Report>  list = new ArrayList<>() ;
+
+		ArrayList<Report> list = new ArrayList<>();
 
 		ResultSet rset = null;
 		PreparedStatement pstmt = null;
@@ -94,13 +93,13 @@ public class ReportDao {
 
 		if (search.equals("userId")) {
 			sql = sql.replaceAll("##", "USER_ID");
-		} else if(search.equals("receveId")) {
+		} else if (search.equals("receveId")) {
 			sql = sql.replaceAll("##", "RECEIVE_ID");
 		} else {
 			sql = sql.replaceAll("##", "REPORT_TITLE");
 		}
-		
-		// like문에 들어갈 문자열 혹시 모를 _ 와 % escape 처리 
+
+		// like문에 들어갈 문자열 혹시 모를 _ 와 % escape 처리
 		String like = keyword.replaceAll("%", "\\%");
 		like = like.replaceAll("_", "\\_");
 		like = "%" + like + "%";
@@ -110,25 +109,20 @@ public class ReportDao {
 
 			pstmt.setString(1, keyword);
 			pstmt.setString(2, like);
-			
+
 			// 한페이지에 20개
-			pstmt.setInt(3, (page-1)*20 +1);
-			pstmt.setInt(4, page*20);
+			pstmt.setInt(3, (page - 1) * 20 + 1);
+			pstmt.setInt(4, page * 20);
 
 			rset = pstmt.executeQuery();
-			
-			//date 포맷용
-			DateFormat df = new SimpleDateFormat("yy/MM/dd");  
+
+			// date 포맷용
+			DateFormat df = new SimpleDateFormat("yy/MM/dd");
 
 			while (rset.next()) {
-				Report r = new Report(
-							rset.getInt("REPORT_NO"),
-							rset.getString("USER_ID"),
-							rset.getString("RECEIVE_ID"),
-							rset.getString("REPORT_TITLE"),
-							rset.getString("REPORT_CONTENT"),
-							df.format(rset.getDate("CREATE_DATE"))
-						);
+				Report r = new Report(rset.getInt("REPORT_NO"), rset.getString("USER_ID"), rset.getString("RECEIVE_ID"),
+						rset.getString("REPORT_TITLE"), rset.getString("REPORT_CONTENT"),
+						df.format(rset.getDate("CREATE_DATE")));
 				list.add(r);
 			}
 		} catch (SQLException e) {
@@ -137,20 +131,18 @@ public class ReportDao {
 			close(rset);
 			close(pstmt);
 		}
-		
+
 		return list;
 	}
-	
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public Report selectReport(Connection conn, int reportNo) {
-		Report r = null ;
+		Report r = null;
 
 		ResultSet rset = null;
 		PreparedStatement pstmt = null;
 
 		String sql = prop.getProperty("selectReport");
-
 
 		try {
 			pstmt = conn.prepareStatement(sql);
@@ -158,19 +150,14 @@ public class ReportDao {
 			pstmt.setInt(1, reportNo);
 
 			rset = pstmt.executeQuery();
-			
-			//date 포맷용
-			DateFormat df = new SimpleDateFormat("yy/MM/dd HH:dd");  
+
+			// date 포맷용
+			DateFormat df = new SimpleDateFormat("yy/MM/dd HH:dd");
 
 			if (rset.next()) {
-				r = new Report(
-							rset.getInt("REPORT_NO"),
-							rset.getString("USER_ID"),
-							rset.getString("RECEIVE_ID"),
-							rset.getString("REPORT_TITLE"),
-							rset.getString("REPORT_CONTENT"),
-							df.format(rset.getTimestamp("CREATE_DATE"))
-						);
+				r = new Report(rset.getInt("REPORT_NO"), rset.getString("USER_ID"), rset.getString("RECEIVE_ID"),
+						rset.getString("REPORT_TITLE"), rset.getString("REPORT_CONTENT"),
+						df.format(rset.getTimestamp("CREATE_DATE")));
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -178,20 +165,18 @@ public class ReportDao {
 			close(rset);
 			close(pstmt);
 		}
-		
+
 		return r;
 	}
-	
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public int selectUserCount(Connection conn, String userId) {
 		int result = 0;
-		
+
 		ResultSet rset = null;
 		PreparedStatement pstmt = null;
 
 		String sql = prop.getProperty("selectUserCount");
-
 
 		try {
 			pstmt = conn.prepareStatement(sql);
@@ -210,14 +195,14 @@ public class ReportDao {
 			close(rset);
 			close(pstmt);
 		}
-		
+
 		return result;
 	}
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public int selectReceiverCount(Connection conn, String receiveId) {
 		int result = 0;
-		
+
 		ResultSet rset = null;
 		PreparedStatement pstmt = null;
 
@@ -240,19 +225,16 @@ public class ReportDao {
 			close(rset);
 			close(pstmt);
 		}
-		
+
 		return result;
 	}
-	
-	
-	
+
 //------------------------------ update 구간 -------------------------------
-	
-	
-	//[han] 어드민페이지용
+
+	// [han] 어드민페이지용
 	public int updateReport(Connection conn, Report report) {
-		int result= 0;
-		
+		int result = 0;
+
 		PreparedStatement pstmt = null;
 
 		String sql = prop.getProperty("updateReport");
@@ -270,21 +252,23 @@ public class ReportDao {
 		} finally {
 			close(pstmt);
 		}
-		
+
 		return result;
 	}
 	
-	//[han] 어드민페이지용
+//------------------------------ insert 구간 -------------------------------
+
+	// [han] 어드민페이지용
 	public int deleteReport(Connection conn, int reportNo) {
-		int result= 0;
-		
+		int result = 0;
+
 		PreparedStatement pstmt = null;
 
 		String sql = prop.getProperty("deleteReport");
 
 		try {
 			pstmt = conn.prepareStatement(sql);
-			
+
 			pstmt.setInt(1, reportNo);
 
 			result = pstmt.executeUpdate();
@@ -293,29 +277,35 @@ public class ReportDao {
 		} finally {
 			close(pstmt);
 		}
-		
+
 		return result;
 	}
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
+
+	// 신고 정보 db 저장 - 가영
+	public int insertReport(Connection conn, String userId, String receiveId, String reportTitle,
+			String reportContent) {
+
+		int result = 0;
+
+		PreparedStatement pstmt = null;
+
+		String sql = prop.getProperty("insertReport");
+
+		try {
+			pstmt = conn.prepareStatement(sql);
+
+			pstmt.setString(1, userId);
+			pstmt.setString(2, receiveId);
+			pstmt.setString(3, reportTitle);
+			pstmt.setString(4, reportContent);
+
+			result = pstmt.executeUpdate();
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} finally {
+			close(pstmt);
+		}
+		return result;
+	}
+
 }
