+++ SEMI_MZ/src/sql/member/member-mapper.xml
@@ -161,20 +161,6 @@
 		WHERE USER_ID = ?
 	</entry>
 	
-	<!-- [가영] 신고 정보 db 저장 -->
-	<entry key="insertReport">
-		INSERT INTO REPORT (REPORT_NO,
-							USER_ID,
-							RECEIVE_ID,
-							REPORT_TITLE,
-							REPORT_CONTENT)
-		VALUES (SEQ_REPORT.NEXTVAL,
-				?,
-				?,
-				?,
-				?)
-	</entry>
-	
 	<!-- [가영] 호감도 추가, 삭제, 조회 -->
 	<entry key="insertHeart">
 		INSERT INTO HEART(USER_ID, RECEIVE_ID)
